Struct fdt::Fdt

pub struct Fdt<'dtb> {
    pub memory_reservations: &'dtb [MemoryReservation],
    /* private fields */
}

Flattened Device Tree

Fields
memory_reservations: &'dtb [MemoryReservation]

Memory Reservation Block

Implementations

impl<'dtb> Fdt<'dtb>

pub fn new(dtb: &'dtb [u8]) -> Result<Fdt<'dtb>, Error>

Create an Fdt from a buffer

Errors

This function

pub unsafe fn from_ptr(ptr: *const u8) -> Result<Fdt<'dtb>, Error>

Create an Fdt from a raw pointer

Safety

The pointer must point to a valid device tree blob. Specifically, it must be at least 8-byte aligned and valid for reads up to the number of bytes specified in the header.

Errors

This function errors for the same reasons as Fdt::new().

pub const fn as_ptr(&self) -> *const u8

Returns a pointer to the underlying device tree blob

pub fn total_size(&self) -> usize

Returns the total size of the device tree blob, in bytes

pub fn get_string(&self, offset: u32) -> Option<&'dtb str>

Returns the string at the specified offset in the string table

pub fn root(&self) -> Node<'_, 'dtb>

pub fn try_find_node(&self, path: &str) -> Result<Option<Node<'_, 'dtb>>>

pub fn find_node(&self, path: &str) -> Option<Node<'_, 'dtb>>

pub fn cpus(&self) -> impl Iterator<Item = Node<'_, 'dtb>>

pub fn try_property_as<'f, T: Parse<'f, 'dtb>>( &'f self, path: &str ) -> Result<Option<T>>where 'dtb: 'f,

pub fn property_as<'f, T: Parse<'f, 'dtb>>(&'f self, path: &str) -> Option<T>where 'dtb: 'f,